Disable best-effort escape by default, fixes #278.
The default strategy exits instead.
