+++ core/src/test/resources/default140/Apply.stat
@@ -1,0 +1,13 @@
+// Columns 140                                                                                                                             |
+<<< SKIP idempotent #278
+trait SelfEmploymentsBalancingChargesController extends BaseController with Links {
+
+  val balancingChargesJson = toJson(
+      balancingCharges.map(balancingCharge =>
+            halResource(obj(),
+                        Seq(HalLink("self",
+                                    selfEmploymentSummaryTypeIdHref(
+                                        saUtr, taxYear, seId, BalancingChargesSummaryType, balancingCharge.id.get))))))
+}
+>>>
+x
