+++ core/src/test/scala/org/scalafmt/util/HasTests.scala
@@ -93,6 +93,7 @@ trait HasTests extends FunSuiteLike with FormatAssertions {
     filename.split("/").reverse(1) match {
       case "unit" => ScalafmtStyle.unitTest40
       case "default" | "standard" | "scala" => ScalafmtStyle.unitTest80
+      case "default140" => ScalafmtStyle.default.copy(maxColumn = 140)
       case "scalajs" => ScalafmtStyle.scalaJs
       case "stripMargin" => ScalafmtStyle.default
       case "spaces" =>
